tests/test_phase_g6_ws.py
from __future__ import annotations

import asyncio
import importlib.util
import json
from pathlib import Path
from unittest.mock import AsyncMock, MagicMock, patch

import pytest

from api.websocket_hub import WebSocketHub

MODULE_PATH = Path(__file__).resolve().parent.parent / "basic-blockchain.py"


def _load_module():
    spec = importlib.util.spec_from_file_location("basic_blockchain", MODULE_PATH)
    mod = importlib.util.module_from_spec(spec)
    spec.loader.exec_module(mod)
    return mod


# ---------------------------------------------------------------------------
# WebSocketHub — unit tests
# ---------------------------------------------------------------------------

async def test_hub_starts_with_zero_connections():
    hub = WebSocketHub()
    assert hub.connection_count == 0


async def test_hub_broadcast_delivers_to_connected_queue():
    hub = WebSocketHub()
    q = hub._make_queue()
    hub.broadcast({"event": "block_mined", "block": {"index": 2}})
    message = q.get_nowait()
    assert json.loads(message) == {"event": "block_mined", "block": {"index": 2}}
    hub._remove_queue(q)


async def test_hub_broadcast_delivers_to_multiple_queues():
    hub = WebSocketHub()
    q1 = hub._make_queue()
    q2 = hub._make_queue()
    hub.broadcast({"event": "ping"})
    assert not q1.empty()
    assert not q2.empty()
    hub._remove_queue(q1)
    hub._remove_queue(q2)


async def test_hub_broadcast_skips_when_no_clients():
    hub = WebSocketHub()
    hub.broadcast({"event": "ping"})  # must not raise


async def test_hub_remove_queue_decrements_count():
    hub = WebSocketHub()
    q = hub._make_queue()
    assert hub.connection_count == 1
    hub._remove_queue(q)
    assert hub.connection_count == 0


async def test_hub_serve_sends_message_and_exits_on_cancel():
    hub = WebSocketHub()
    sent = []

    async def fake_send(msg):
        sent.append(msg)

    async def driver():
        task = asyncio.create_task(hub.serve(send_fn=fake_send))
        await asyncio.sleep(0)
        hub.broadcast({"event": "block_mined"})
        await asyncio.sleep(0.05)
        task.cancel()
        try:
            await task
        except asyncio.CancelledError:
            pass

    await driver()

    assert len(sent) == 1
    assert json.loads(sent[0]) == {"event": "block_mined"}
    assert hub.connection_count == 0


async def test_hub_connection_count_tracks_active_serve_calls():
    hub = WebSocketHub()
    assert hub.connection_count == 0
    q = hub._make_queue()
    assert hub.connection_count == 1
    q2 = hub._make_queue()
    assert hub.connection_count == 2
    hub._remove_queue(q)
    hub._remove_queue(q2)
    assert hub.connection_count == 0


# ---------------------------------------------------------------------------
# API — mine_block broadcasts to WebSocket hub
# ---------------------------------------------------------------------------

async def test_mine_block_calls_hub_broadcast():
    mod = _load_module()
    hub = WebSocketHub()
    app = mod.create_app(ws_hub=hub)
    broadcasts = []
    original_broadcast = hub.broadcast

    def capture(payload):
        broadcasts.append(payload)
        original_broadcast(payload)

    hub.broadcast = capture

    async with app.test_client() as client:
        with patch("domain.PropagationService.notify_resolve"):
            await client.post("/api/v1/mine_block")

    assert len(broadcasts) == 1
    assert broadcasts[0]["event"] == "block_mined"
    assert "block" in broadcasts[0]


async def test_ws_endpoint_reachable():
    mod = _load_module()
    app = mod.create_app()
    async with app.test_client() as client:
        async with client.websocket("/api/v1/ws") as ws:
            # connection established — send a broadcast and receive it
            mod_hub = app.config.get("hub")  # hub is closure-bound; test via mine
            pass  # connection opened without error is sufficient
