requirements.txt
quart>=0.19,<1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
PyJWT>=2.8
bcrypt>=4.1
mnemonic>=0.21
coincurve>=20.0
